refactor: ApplyGuide styles
