+++ src/pages/ApplyPage/components/Guide/ApplyGuide.js
@@ -6,14 +6,11 @@ import Banner from "../../../../components/Layouts/Banner";
 import Culture from "./Culture";
 import ApplyProgress from "components/ApplyProgress";
 
-import ApplyGuide_2 from "./ApplyGuide_2";
-
 const ApplyGuide = () => {
   return (
     <>
       <Banner title="지원안내" />"
       <Culture />
-      {/* <ApplyGuide_2 /> */}
       <ApplyProgress />
     </>
   );
