+++ src/pages/ApplyPage/components/styles/CompleteMessage.css
