+++ src/pages/ApplyPage/components/styles/ExistenceApplication.css
